p4i integration Integrate 214575 from 4.1 to main: Fixes:    Designer crash on 64-bit platforms with 32-bit binaries Task:     107605 RevBy:    Brad Details:  QModelIndex stores internalId() as a void*. Pointers are 	  usually 64-bit wide, but in 32-bit binaries they are                  still 32-bits. So when I would retrieve a value of -1,                  a 32-bit pointer would be cast to a 64-bit int, and                  would not be -1 any more.
[git-p4: depot-paths = "//depot/qt/main/": change = 214576]

+++ tools/designer/src/lib/shared/resourcefile.cpp
@@ -384,7 +384,7 @@ QModelIndex ResourceModel::index(int row, int column,
 {
     QModelIndex result;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid()) {
         if (row < m_resource_file.prefixCount())
@@ -403,7 +403,7 @@ QModelIndex ResourceModel::parent(const QModelIndex &index) const
 {
     QModelIndex result;
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     if (index.isValid() && d != -1)
         result = createIndex(d, 0, -1);
@@ -415,7 +415,7 @@ int ResourceModel::rowCount(const QModelIndex &parent) const
 {
     int result = 0;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid())
         result = m_resource_file.prefixCount();
@@ -434,7 +434,7 @@ bool ResourceModel::hasChildren(const QModelIndex &parent) const
 {
     bool result = false;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid())
         result = m_resource_file.prefixCount() > 0;
@@ -466,7 +466,7 @@ QVariant ResourceModel::data(const QModelIndex &index, int role) const
     if (!index.isValid())
         return QVariant();
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     QVariant result;
 
@@ -509,7 +509,7 @@ void ResourceModel::getItem(const QModelIndex &index, QString &prefix, QString &
     if (!index.isValid())
         return;
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     if (d == -1) {
         prefix = m_resource_file.prefix(index.row());
